Ethernet: receive connectivity changes
Previous implementation - EthernetBootReceiver - was only created along the
Activity and do not handle changes in the background.
New implementation - EthernetReceive - is now started at bootup and
stays in background.
