+++ Ethernet/AndroidManifest.xml
@@ -28,11 +28,12 @@
             </intent-filter>
         </activity>
         <receiver
-            android:name=".EthernetBootReceiver"
+            android:name=".EthernetReceiver"
             android:enabled="true"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.BOOT_COMPLETED" />
+                <action android:name="android.net.conn.CONNECTIVITY_CHANGE" />
             </intent-filter>
         </receiver>
     </application>
