--- Ethernet/com/fsl/ethernet/EthernetBootReceiver.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2015 Boundary Devices, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.fsl.ethernet;
-
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-import android.util.Log;
-import com.fsl.ethernet.EthernetManager;
-
-public class EthernetBootReceiver extends BroadcastReceiver {
-
-    private static final String TAG = "EthernetBootReceiver";
-
-    @Override
-    public void onReceive(Context context, Intent intent) {
-        Log.d(TAG, "onReceive " + intent.getAction());
-        if (Intent.ACTION_BOOT_COMPLETED.equals(intent.getAction())) {
-            Log.d(TAG, "resetting interface");
-            EthernetManager ethManager = new EthernetManager(context);
-            ethManager.resetInterface();
-        }
-    }
-}
