+++ Ethernet/com/fsl/ethernet/MainActivity.java
@@ -54,25 +54,6 @@ public class MainActivity extends Activity {
     public static final String FIRST_RUN = "ethernet";
     private Button mBtnAdvanced;
     private EthernetAdvDialog mEthAdvancedDialog;
-    private final BroadcastReceiver mEthernetReceiver = new BroadcastReceiver() {
-        public void onReceive(Context context, Intent intent) {
-            if(ConnectivityManager.CONNECTIVITY_ACTION.equals(intent.getAction())){
-                NetworkInfo info =
-                    intent.getParcelableExtra(ConnectivityManager.EXTRA_NETWORK_INFO);
-                if (info != null) {
-                    Log.i(TAG,"getState()="+info.getState() + "getType()=" +
-                            info.getType());
-                    if (info.getType() == ConnectivityManager.TYPE_ETHERNET) {
-                        if (info.getState() == State.DISCONNECTED)
-                            mConnMgr.setGlobalProxy(null);
-                            SystemProperties.set("rw.HTTP_PROXY", "");
-                        if (info.getState() == State.CONNECTED)
-                            mEthEnabler.getManager().initProxy();
-                    }
-                }
-            }
-        }
-    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -88,10 +69,8 @@ public class MainActivity extends Activity {
         addListenerOnBtnConfig();
         addListenerOnBtnCheck();
         addListenerOnBtnAdvanced();
-        mConnMgr = (ConnectivityManager)getSystemService(Context.CONNECTIVITY_SERVICE);
         IntentFilter filter = new IntentFilter();
         filter.addAction(ConnectivityManager.CONNECTIVITY_ACTION);
-        registerReceiver(mEthernetReceiver, filter);
     }
 
 
@@ -171,6 +150,5 @@ public class MainActivity extends Activity {
     public void onDestroy() {
         super.onDestroy();
         Log.i(TAG, "onStop() will force clear global proxy set by ethernet");
-        unregisterReceiver(mEthernetReceiver);
     }
 }
